+++ src/lib/components/CoverCard.svelte
@@ -19,22 +19,12 @@
   <div class="albums">
     <div class="album">
       {#if original?.album_img}
-        <img
-          data-src={original.album_img[1]}
-          src={original.album_img[2]}
-          alt={`${original.name} album art`}
-          class="lazyload"
-        />
+        <img src={original.album_img[1]} alt={`${original.name} album art`} />
       {/if}
     </div>
     <div class="album">
       {#if cover?.album_img}
-        <img
-          data-src={cover.album_img[1]}
-          src={cover.album_img[2]}
-          alt={`${cover.name} album art`}
-          class="lazyload"
-        />
+        <img src={cover.album_img[1]} alt={`${cover.name} album art`} />
       {/if}
     </div>
   </div>
